+++ resources/[dg]/dg-gangs/typescript/src/server/server.ts
@@ -3,6 +3,7 @@ import { dispatchCurrentGangToAllClients } from 'helpers';
 
 import './controllers';
 import './services/core';
+import './services/keys';
 
 setImmediate(async () => {
   await gangManager.loadAllGangs();